+++ src/schemas/questionCondition.ts
@@ -55,7 +55,7 @@ export const typeDefs = gql`
     "The type of condition/operator to evaluate"
     conditionType: QuestionConditionCondition!
     "The value(s) to match on"
-    conditionMatch: String
+    conditionMatch: String!
   }
 
 `